third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/shared-workers.https-expected.txt

This is a testharness.js-based test.
[FAIL] require-corp (derived from blob URL creator)
  assert_equals: expected "failure" but got "success"
[FAIL] "require-corp" (derived from owner set due to use of local scheme - blob URL)
  assert_equals: expected "failure" but got "success"
[FAIL] "require-corp" (derived from owner set due to use of local scheme - data URL)
  assert_equals: expected "failure" but got "success"
[FAIL] default policy (derived from owner set due to use of local scheme - filesystem URL)
  assert_unreached: SharedWorker.onerror should not be called Reached unreachable code
[FAIL] require-corp (derived from filesystem URL creator)
  assert_unreached: SharedWorker.onerror should not be called Reached unreachable code
[FAIL] "require-corp" (derived from owner set due to use of local scheme - filesystem URL)
  assert_unreached: SharedWorker.onerror should not be called Reached unreachable code
Harness: the test ran to completion.
